src/constants/index.ts: TAB_LABELS
