+++ src/constants/index.ts
@@ -1,0 +1,6 @@
+export const TAB_LABELS = {
+  PROFILE: 'Profile',
+  SUMMARY: 'Summary',
+  AUTO_COLLECTIONS: 'Updates',
+  COLLECTIONS: 'Collections',
+};
